move methods to sia module
